@@ -1,41 +1,50 @@
 {
   "name": "catalog",
-  "version": "2.3.1",
+  "version": "2.4.0",
   "private": true,
   "scripts": {
     "android": "react-native run-android",
     "ios": "react-native run-ios",
     "start": "react-native start",
     "test": "jest",
-    "lint": "eslint --ext .js,.jsx,.ts,.tsx ./helpers ./examples ./configuration",
+    "lint": "eslint --ext .js,.jsx,.ts,.tsx ./*.js __tests__/*.js ./helpers ./examples ./configuration",
     "postinstall": "rm -rf ./node_modules/react-native-pspdfkit/samples/*/node_modules"
   },
   "dependencies": {
-    "@babel/preset-env": "^7.18.6",
-    "@react-native-masked-view/masked-view": "^0.2.6",
+    "@babel/preset-env": "^7.19.4",
+    "@react-native-masked-view/masked-view": "^0.2.8",
+    "@react-navigation/native": "^6.0.13",
+    "@react-navigation/native-stack": "^6.9.1",
     "prop-types": "^15.8.1",
     "react": "^18.2.0",
-    "react-native": "^0.69.1",
+    "react-native": "^0.70.3",
     "react-native-fs": "^2.20.0",
-    "react-native-gesture-handler": "^2.5.0",
+    "react-native-gesture-handler": "^2.8.0",
     "react-native-pspdfkit": "file:../..",
-    "react-native-safe-area-context": "^4.3.1",
-    "react-native-screens": "^3.15.0",
-    "react-native-windows": "^0.69.1",
+    "react-native-safe-area-context": "^4.4.1",
+    "react-native-screens": "^3.18.2",
+    "react-native-windows": "^0.70.4",
     "react-navigation": "^4.4.4",
     "react-navigation-stack": "^2.10.4",
-    "typescript": "^4.7.4"
+    "typescript": "^4.8.4",
+    "@xmldom/xmldom": ">=0.8.3"
   },
   "devDependencies": {
-    "@babel/core": "^7.18.6",
-    "@babel/runtime": "^7.18.6",
-    "@react-native-community/eslint-config": "^3.0.3",
-    "babel-jest": "^28.1.2",
-    "eslint": "^8.19.0",
-    "jest": "^28.1.2",
-    "metro-react-native-babel-preset": "^0.71.3",
-    "react-native-codegen": "^0.69.1",
-    "react-test-renderer": "18.2.0"
+    "@babel/core": "^7.19.6",
+    "@babel/eslint-parser": "^7.19.1",
+    "@babel/runtime": "^7.19.4",
+    "@react-native-community/eslint-config": "^3.1.0",
+    "babel-jest": "^29.2.2",
+    "eslint": "^8.26.0",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-plugin-prettier": "^4.2.1",
+    "eslint-plugin-react": "^7.31.10",
+    "jest": "^29.2.2",
+    "metro-react-native-babel-preset": "^0.73.2",
+    "prettier": "^2.7.1",
+    "react-native-codegen": "^0.70.6",
+    "react-test-renderer": "18.2.0",
+    "eslint-plugin-ft-flow":"^2.0.1"
   },
   "jest": {
     "preset": "react-native",
