@@ -2,12 +2,12 @@
  * @format
  */
 
-import 'react-native';
-import React from 'react';
+// import React from 'react';
 
 // Note: test renderer must be required after react-native.
-import renderer from 'react-test-renderer';
+// import renderer from 'react-test-renderer';
+// import Catalog from '../Catalog';
 
 it('renders correctly', () => {
-  // Add test here
+  // add tests here
 });
