@@ -8,26 +8,51 @@ import {
 import React from 'react';
 
 export class ManualSave extends BaseExampleAutoHidingHeaderComponent {
+  pdfRef = null;
+
+  constructor(props) {
+    super(props);
+    const { navigation } = this.props;
+    this.pdfRef = React.createRef();
+
+    navigation.addListener('beforeRemove', e => {
+      this.pdfRef?.current?.destroyView();
+    });
+  }
+
+  componentWillUnmount() {
+    const { navigation } = this.props;
+    navigation.removeListener('beforeRemove');
+  }
+
   render() {
     return (
       <View style={styles.flex}>
         <PSPDFKitView
-          ref="pdfView"
+          ref={this.pdfRef}
           document={writableDocumentPath}
           disableAutomaticSaving={true}
           configuration={{
             backgroundColor: processColor('lightgrey'),
           }}
+          menuItemGrouping={[
+            'pen',
+            'freetext',
+            { key: 'markup', items: ['highlight', 'underline'] },
+            'image',
+          ]}
           pageIndex={3}
           style={styles.pdfColor}
         />
         <View style={styles.wrapper}>
           <View style={styles.flex}>
             <Button
+              accessibilityLabel={'Save Button'}
+              testID={'Save Button'}
               onPress={() => {
                 // Manual Save
-                this.refs.pdfView
-                  .saveCurrentDocument()
+                this.pdfRef?.current
+                  ?.saveCurrentDocument()
                   .then(saved => {
                     if (saved) {
                       alert('Successfully saved current document.');
