@@ -3,13 +3,28 @@ import { Button, processColor, View } from 'react-native';
 import PSPDFKitView from 'react-native-pspdfkit';
 import { formDocumentPath, pspdfkitColor } from '../configuration/Constants';
 import React from 'react';
-
 export class ProgrammaticFormFilling extends BaseExampleAutoHidingHeaderComponent {
+  pdfRef = null;
+
+  constructor(props) {
+    super(props);
+    const { navigation } = this.props;
+    this.pdfRef = React.createRef();
+
+    navigation.addListener('beforeRemove', e => {
+      this.pdfRef?.current?.destroyView();
+    });
+  }
+
+  componentWillUnmount() {
+    const { navigation } = this.props;
+    navigation.removeListener('beforeRemove');
+  }
   render() {
     return (
       <View style={styles.flex}>
         <PSPDFKitView
-          ref="pdfView"
+          ref={this.pdfRef}
           document={formDocumentPath}
           disableAutomaticSaving={true}
           configuration={{
@@ -22,7 +37,7 @@ export class ProgrammaticFormFilling extends BaseExampleAutoHidingHeaderComponen
             <Button
               onPress={() => {
                 // Fill Text Form Fields.
-                this.refs.pdfView
+                this.pdfRef.current
                   .setFormFieldValue('Name_Last', 'Appleseed')
                   .then(result => {
                     if (result) {
@@ -34,7 +49,7 @@ export class ProgrammaticFormFilling extends BaseExampleAutoHidingHeaderComponen
                   .catch(error => {
                     alert(JSON.stringify(error));
                   });
-                this.refs.pdfView
+                this.pdfRef.current
                   .setFormFieldValue('Name_First', 'John')
                   .then(result => {
                     if (result) {
@@ -46,7 +61,7 @@ export class ProgrammaticFormFilling extends BaseExampleAutoHidingHeaderComponen
                   .catch(error => {
                     alert(JSON.stringify(error));
                   });
-                this.refs.pdfView
+                this.pdfRef.current
                   .setFormFieldValue('Address_1', '1 Infinite Loop')
                   .then(result => {
                     if (result) {
@@ -58,7 +73,7 @@ export class ProgrammaticFormFilling extends BaseExampleAutoHidingHeaderComponen
                   .catch(error => {
                     alert(JSON.stringify(error));
                   });
-                this.refs.pdfView
+                this.pdfRef.current
                   .setFormFieldValue('City', 'Cupertino')
                   .then(result => {
                     if (result) {
@@ -70,7 +85,7 @@ export class ProgrammaticFormFilling extends BaseExampleAutoHidingHeaderComponen
                   .catch(error => {
                     alert(JSON.stringify(error));
                   });
-                this.refs.pdfView
+                this.pdfRef.current
                   .setFormFieldValue('STATE', 'CA')
                   .then(result => {
                     if (result) {
@@ -82,7 +97,7 @@ export class ProgrammaticFormFilling extends BaseExampleAutoHidingHeaderComponen
                   .catch(error => {
                     alert(JSON.stringify(error));
                   });
-                this.refs.pdfView
+                this.pdfRef.current
                   .setFormFieldValue('SSN', '123456789')
                   .then(result => {
                     if (result) {
@@ -94,7 +109,7 @@ export class ProgrammaticFormFilling extends BaseExampleAutoHidingHeaderComponen
                   .catch(error => {
                     alert(JSON.stringify(error));
                   });
-                this.refs.pdfView
+                this.pdfRef.current
                   .setFormFieldValue('Telephone_Home', '(123) 456-7890')
                   .then(result => {
                     if (result) {
@@ -106,7 +121,7 @@ export class ProgrammaticFormFilling extends BaseExampleAutoHidingHeaderComponen
                   .catch(error => {
                     alert(JSON.stringify(error));
                   });
-                this.refs.pdfView
+                this.pdfRef.current
                   .setFormFieldValue('Birthdate', '1/1/1983')
                   .then(result => {
                     if (result) {
@@ -120,7 +135,7 @@ export class ProgrammaticFormFilling extends BaseExampleAutoHidingHeaderComponen
                   });
 
                 // Select a button form elements.
-                this.refs.pdfView
+                this.pdfRef.current
                   .setFormFieldValue('Sex.0', 'selected')
                   .then(result => {
                     if (result) {
@@ -132,7 +147,7 @@ export class ProgrammaticFormFilling extends BaseExampleAutoHidingHeaderComponen
                   .catch(error => {
                     alert(JSON.stringify(error));
                   });
-                this.refs.pdfView
+                this.pdfRef.current
                   .setFormFieldValue('PHD', 'selected')
                   .then(result => {
                     if (result) {
@@ -146,6 +161,7 @@ export class ProgrammaticFormFilling extends BaseExampleAutoHidingHeaderComponen
                   });
               }}
               title="Fill Form"
+              accessibilityLabel="Fill Form"
             />
           </View>
           <View style={styles.marginLeft}>
@@ -153,10 +169,11 @@ export class ProgrammaticFormFilling extends BaseExampleAutoHidingHeaderComponen
               onPress={async () => {
                 // Get the First Name Value.
                 const firstNameValue =
-                  await this.refs.pdfView.getFormFieldValue('Name_Last');
+                  await this.pdfRef.current.getFormFieldValue('Name_Last');
                 alert(JSON.stringify(firstNameValue));
               }}
               title="Get Last Name Value"
+              accessibilityLabel="Get Last Name Value"
             />
           </View>
         </View>
