@@ -2,12 +2,14 @@ import { Component } from 'react';
 import { Platform } from 'react-native';
 
 export class BaseExampleAutoHidingHeaderComponent extends Component {
-  static navigationOptions = () => {
+  constructor(props) {
+    super(props);
+    const { navigation } = this.props;
     if (Platform.OS === 'android') {
-      return {
+      navigation.setOptions({
         // Since the PSPDFKitView provides it's own toolbar and back button we don't need a header in Android.
         headerShown: false,
-      };
+      });
     }
-  };
+  }
 }
