@@ -2,12 +2,12 @@
  * @format
  */
 
-import 'react-native';
-import React from 'react';
-
+// import React from 'react';
 // Note: test renderer must be required after react-native.
-import renderer from 'react-test-renderer';
+// import renderer from 'react-test-renderer';
+// import NativeCatalog from '../NativeCatalog';
 
 it('renders correctly', () => {
-  // Add test here
+  // const rendered = renderer.create(<NativeCatalog />).toJSON();
+  // expect(rendered).toBeTruthy();
 });
