@@ -1,5 +1,6 @@
 import { BaseExampleAutoHidingHeaderComponent } from '../helpers/BaseExampleAutoHidingHeaderComponent';
 import { CustomPdfView, formDocumentPath } from '../configuration/Constants';
+import { withNavigation } from 'react-navigation';
 import {
   Button,
   findNodeHandle,
@@ -11,27 +12,53 @@ import {
 } from 'react-native';
 import React from 'react';
 
-export class ManualSigning extends BaseExampleAutoHidingHeaderComponent {
+class ManualSigning extends BaseExampleAutoHidingHeaderComponent {
+  pdfRef = React.createRef();
+
   constructor(props) {
     super(props);
+    const { navigation } = props;
     this.state = {
       documentPath: formDocumentPath,
+      shouldReturn: false,
     };
+
+    navigation.addListener('beforeRemove', e => {
+      if (Platform.OS !== 'android') {
+        return;
+      }
+      const { shouldReturn } = this.state;
+      if (!shouldReturn) {
+        this.setState({ shouldReturn: true });
+        e.preventDefault();
+      }
+      setTimeout(() => {
+        this.goBack();
+      }, 50);
+    });
+  }
+
+  goBack(navigation) {
+    this.props.navigation.goBack();
   }
 
   render() {
+    const { shouldReturn } = this.state;
+
     return (
       <View style={styles.flex}>
-        <CustomPdfView
-          ref="pdfView"
-          document={this.state.documentPath}
-          style={styles.flex}
-          onDocumentDigitallySigned={event => {
-            this.setState({
-              documentPath: event.nativeEvent.signedDocumentPath,
-            });
-          }}
-        />
+        {!shouldReturn && (
+          <CustomPdfView
+            ref={this.pdfRef}
+            document={this.state.documentPath}
+            style={styles.flex}
+            onDocumentDigitallySigned={event => {
+              this.setState({
+                documentPath: event.nativeEvent.signedDocumentPath,
+              });
+            }}
+          />
+        )}
         <SafeAreaView style={styles.row}>
           <View>
             <Button
@@ -57,6 +84,8 @@ export class ManualSigning extends BaseExampleAutoHidingHeaderComponent {
     );
   }
 }
+
+export default withNavigation(ManualSigning);
 
 const styles = {
   flex: { flex: 1 },
