@@ -1,7 +1,10 @@
 //
 //  Use this file to import your target's public headers that you would like to expose to Swift.
 //
-
 @import UIKit;
 @import PSPDFKit;
 @import PSPDFKitUI;
+
+#import <React/RCTBridgeModule.h>
+#import <React/RCTEventEmitter.h>
+
