@@ -5,7 +5,7 @@
  * @format
  */
 
- module.exports = {
+module.exports = {
   transformer: {
     getTransformOptions: async () => ({
       transform: {
